pyproject.toml
[tool.poetry]
name = "altonlineab-milestone-4"
version = "0.1.0"
description = ""
authors = ["johnny.lazo <johnny.lazo@cgi.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
mimerpy = "^1.3.7"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
